requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
albucore==0.0.24
    # via albumentations
albumentations==2.0.6
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
attrs==25.3.0
    # via cvat-sdk
bleach==6.2.0
    # via kaggle
certifi==2025.4.26
    # via
    #   kaggle
    #   requests
    #   roboflow
    #   tuspy
charset-normalizer==3.4.2
    # via
    #   kaggle
    #   requests
contourpy==1.3.2
    # via matplotlib
cvat-sdk==2.36.0
    # via -r requirements.in
cycler==0.12.1
    # via
    #   matplotlib
    #   roboflow
filelock==3.18.0
    # via torch
filetype==1.2.0
    # via roboflow
fonttools==4.58.0
    # via matplotlib
fsspec==2025.3.2
    # via torch
future==1.0.0
    # via tuspy
idna==3.7
    # via
    #   kaggle
    #   requests
    #   roboflow
jinja2==3.1.6
    # via torch
joblib==1.5.0
    # via scikit-learn
kaggle==1.7.4.5
    # via -r requirements.in
kiwisolver==1.4.8
    # via
    #   matplotlib
    #   roboflow
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.3
    # via
    #   roboflow
    #   seaborn
    #   ultralytics
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.2.5
    # via
    #   albucore
    #   albumentations
    #   contourpy
    #   matplotlib
    #   opencv-python
    #   opencv-python-headless
    #   pandas
    #   roboflow
    #   scikit-learn
    #   scipy
    #   seaborn
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
nvidia-cublas-cu12==12.6.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.6.80
    # via torch
nvidia-cuda-nvrtc-cu12==12.6.77
    # via torch
nvidia-cuda-runtime-cu12==12.6.77
    # via torch
nvidia-cudnn-cu12==9.5.1.17
    # via torch
nvidia-cufft-cu12==11.3.0.4
    # via torch
nvidia-cufile-cu12==1.11.1.6
    # via torch
nvidia-curand-cu12==10.3.7.77
    # via torch
nvidia-cusolver-cu12==11.7.1.2
    # via torch
nvidia-cusparse-cu12==12.5.4.2
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.3
    # via torch
nvidia-nccl-cu12==2.26.2
    # via torch
nvidia-nvjitlink-cu12==12.6.85
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.6.77
    # via torch
opencv-python==4.11.0.86
    # via
    #   -r requirements.in
    #   ultralytics
opencv-python-headless==4.10.0.84
    # via
    #   albucore
    #   albumentations
    #   roboflow
packaging==25.0
    # via
    #   cvat-sdk
    #   matplotlib
pandas==2.2.3
    # via
    #   seaborn
    #   ultralytics
pillow==11.2.1
    # via
    #   cvat-sdk
    #   matplotlib
    #   pillow-heif
    #   roboflow
    #   torchvision
    #   ultralytics
pillow-heif==0.22.0
    # via roboflow
platformdirs==4.3.8
    # via cvat-sdk
protobuf==6.30.2
    # via kaggle
psutil==7.0.0
    # via ultralytics
py-cpuinfo==9.0.0
    # via ultralytics
pydantic==2.11.4
    # via albumentations
pydantic-core==2.33.2
    # via pydantic
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   cvat-sdk
    #   kaggle
    #   matplotlib
    #   pandas
    #   roboflow
python-dotenv==1.1.0
    # via roboflow
python-slugify==8.0.4
    # via kaggle
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   albumentations
    #   roboflow
    #   ultralytics
requests==2.32.3
    # via
    #   kaggle
    #   requests-toolbelt
    #   roboflow
    #   tuspy
    #   ultralytics
requests-toolbelt==1.0.0
    # via roboflow
roboflow==1.1.63
    # via -r requirements.in
scikit-learn==1.6.1
    # via -r requirements.in
scipy==1.15.3
    # via
    #   albumentations
    #   scikit-learn
    #   ultralytics
seaborn==0.13.2
    # via ultralytics
simsimd==6.2.1
    # via albucore
six==1.17.0
    # via
    #   kaggle
    #   python-dateutil
    #   roboflow
    #   tuspy
stringzilla==3.12.5
    # via albucore
sympy==1.14.0
    # via torch
text-unidecode==1.3
    # via
    #   kaggle
    #   python-slugify
threadpoolctl==3.6.0
    # via scikit-learn
tinydb==4.8.2
    # via tuspy
torch==2.7.0
    # via
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
torchvision==0.22.0
    # via ultralytics
tqdm==4.67.1
    # via
    #   cvat-sdk
    #   kaggle
    #   roboflow
    #   ultralytics
triton==3.3.0
    # via torch
tuspy==0.2.5
    # via cvat-sdk
typing-extensions==4.13.2
    # via
    #   cvat-sdk
    #   pydantic
    #   pydantic-core
    #   torch
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
tzdata==2025.2
    # via pandas
ultralytics==8.3.131
    # via -r requirements.in
ultralytics-thop==2.0.14
    # via ultralytics
urllib3==2.4.0
    # via
    #   cvat-sdk
    #   kaggle
    #   requests
    #   roboflow
webencodings==0.5.1
    # via
    #   bleach
    #   kaggle

# The following packages are considered to be unsafe in a requirements file:
# setuptools
